Implemented Move Forward
Implemented robot function move forward.

For movinf forward, he checks first if others robots has fallen from the map with that movement. If not, the robot just move forward.

Once the robot has move forward, we check if we lost him. If we lost him, we only need to save the past position to save the limit, and were he was aiming at that moment, because he could be in a limit but pointing into another direction.